requirements.txt
argon2-cffi==23.1.0
pyperclip==1.8.2
cryptography==42.0.2